first commit sprint 3

+++ src/main/resources/static/web/account.js
@@ -50,7 +50,7 @@ const options = {
             return "";
         },
         formattedDate(date) {
-            const options = { year: "numeric", month: "long", day: "numeric" , hour: '2-digit', minute: '2-digit' , second: '2-digit'};
+            const options = { year: "numeric", month: "numeric", day: "numeric" , hour: '2-digit', minute: '2-digit' , second: '2-digit'};
             const formatted = new Date(date).toLocaleDateString(undefined, options);
             return formatted;
         }
